+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.1] - 2023-05-09
+
+### Added
+
+- introduced lock wait timeout into `Database` init
+
+### Fixed
+
+- unlock database after first init
+
 ## [0.1.0] - 2023-05-09
 
 ### Added